+++ sdk/cosmos/azure-cosmos/azure/cosmos/_global_partition_endpoint_manager_circuit_breaker.py
@@ -21,24 +21,17 @@
 
 """Internal class for global endpoint manager for circuit breaker.
 """
-import logging
-import os
 from typing import TYPE_CHECKING
 
-from azure.cosmos import documents
+from azure.cosmos._global_partition_endpoint_manager_circuit_breaker_core import \
+    _GlobalPartitionEndpointManagerForCircuitBreakerCore
 
-from azure.cosmos._partition_health_tracker import PartitionHealthTracker
-from azure.cosmos._routing.routing_range import Range, PartitionKeyRangeWrapper
 from azure.cosmos.aio._global_endpoint_manager_async import _GlobalEndpointManager
-from azure.cosmos._location_cache import EndpointOperationType
 from azure.cosmos._request_object import RequestObject
-from azure.cosmos.http_constants import ResourceType, HttpHeaders
-from azure.cosmos._constants import _Constants as Constants
 if TYPE_CHECKING:
-    from azure.cosmos.aio._cosmos_client_connection_async import CosmosClientConnection
+    from azure.cosmos._cosmos_client_connection import CosmosClientConnection
 
 
-logger = logging.getLogger("azure.cosmos._GlobalEndpointManagerForCircuitBreaker")
 
 class _GlobalPartitionEndpointManagerForCircuitBreaker(_GlobalEndpointManager):
     """
@@ -49,88 +42,33 @@ class _GlobalPartitionEndpointManagerForCircuitBreaker(_GlobalEndpointManager):
 
     def __init__(self, client: "CosmosClientConnection"):
         super(_GlobalPartitionEndpointManagerForCircuitBreaker, self).__init__(client)
-        self.partition_health_tracker = PartitionHealthTracker()
+        self.global_partition_endpoint_manager_core = _GlobalPartitionEndpointManagerForCircuitBreakerCore(client, self.location_cache)
 
     def is_circuit_breaker_applicable(self, request: RequestObject) -> bool:
         """
         Check if circuit breaker is applicable for a request.
         """
-        if not request:
-            return False
-
-        circuit_breaker_enabled = os.environ.get(Constants.CIRCUIT_BREAKER_ENABLED_CONFIG,
-                                                 Constants.CIRCUIT_BREAKER_ENABLED_CONFIG_DEFAULT) == "True"
-        if not circuit_breaker_enabled:
-            return False
-
-        if (not self.location_cache.can_use_multiple_write_locations_for_request(request)
-                and documents._OperationType.IsWriteOperation(request.operation_type)):
-            return False
-
-        if request.resource_type != ResourceType.Document:
-            return False
-
-        if request.operation_type != documents._OperationType.QueryPlan:
-            return False
-
-        return True
-
-    def _create_pkrange_wrapper(self, request: RequestObject) -> PartitionKeyRangeWrapper:
-        """
-        Create a PartitionKeyRangeWrapper object.
-        """
-        container_rid = request.headers[HttpHeaders.IntendedCollectionRID]
-        partition_key = request.headers[HttpHeaders.PartitionKey]
-        # get the partition key range for the given partition key
-        target_container_link = None
-        for container_link, properties in self.client._container_properties_cache:
-            # TODO: @tvaron3 consider moving this to a constant with other usages
-            if properties["_rid"] == container_rid:
-                target_container_link = container_link
-        # throw exception if it is not found
-        pkrange = self.client._routing_map_provider.get_overlapping_range(target_container_link, partition_key)
-        return PartitionKeyRangeWrapper(pkrange, container_rid)
+        return self.global_partition_endpoint_manager_core.is_circuit_breaker_applicable(request)
 
     def record_failure(
             self,
             request: RequestObject
     ) -> None:
-        if self.is_circuit_breaker_applicable(request):
-            #convert operation_type to EndpointOperationType
-            endpoint_operation_type = EndpointOperationType.WriteType if (
-                documents._OperationType.IsWriteOperation(request.operation_type)) else EndpointOperationType.ReadType
-            location = self.location_cache.get_location_from_endpoint(request.location_endpoint_to_route)
-            pkrange_wrapper = self._create_pkrange_wrapper(request)
-            self.partition_health_tracker.add_failure(pkrange_wrapper, endpoint_operation_type, location)
+        self.global_partition_endpoint_manager_core.record_failure(request)
 
     def resolve_service_endpoint(self, request):
-        if self.is_circuit_breaker_applicable(request):
-            pkrange_wrapper = self._create_pkrange_wrapper(request)
-            request.set_excluded_locations_from_circuit_breaker(
-               self.partition_health_tracker.get_excluded_locations(pkrange_wrapper)
-            )
+        request = self.global_partition_endpoint_manager_core.add_excluded_locations_to_request(request)
         return super(_GlobalPartitionEndpointManagerForCircuitBreaker, self).resolve_service_endpoint(request)
-
 
     def mark_partition_unavailable(self, request: RequestObject) -> None:
         """
         Mark the partition unavailable from the given request.
         """
-        location = self.location_cache.get_location_from_endpoint(request.location_endpoint_to_route)
-        pkrange_wrapper = self._create_pkrange_wrapper(request)
-        self.partition_health_tracker.mark_partition_unavailable(pkrange_wrapper, location)
+        self.global_partition_endpoint_manager_core.mark_partition_unavailable(request)
 
     def record_success(
             self,
             request: RequestObject
     ) -> None:
-        if self.is_circuit_breaker_applicable(request):
-            #convert operation_type to either Read or Write
-            endpoint_operation_type = EndpointOperationType.WriteType if (
-                documents._OperationType.IsWriteOperation(request.operation_type)) else EndpointOperationType.ReadType
-            location = self.location_cache.get_location_from_endpoint(request.location_endpoint_to_route)
-            pkrange_wrapper = self._create_pkrange_wrapper(request)
-            self.partition_health_tracker.add_success(pkrange_wrapper, endpoint_operation_type, location)
-
-# TODO: @tvaron3 there should be no in region retries when trying on healthy tentative -----------------------
+        self.global_partition_endpoint_manager_core.record_success(request)