+++ sdk/cosmos/azure-cosmos/azure/cosmos/_request_object.py
@@ -21,7 +21,7 @@
 
 """Represents a request object.
 """
-from typing import Optional, Mapping, Any, List, Dict
+from typing import Optional, Mapping, Any, Dict, Set
 
 
 class RequestObject(object):
@@ -41,7 +41,8 @@ class RequestObject(object):
         self.location_index_to_route: Optional[int] = None
         self.location_endpoint_to_route: Optional[str] = None
         self.last_routed_location_endpoint_within_region: Optional[str] = None
-        self.excluded_locations: Optional[List[str]] = None
+        self.excluded_locations: Optional[Set[str]] = None
+        self.excluded_locations_circuit_breaker: Set[str] = set()
 
     def route_to_location_with_preferred_location_flag(  # pylint: disable=name-too-long
         self,
@@ -83,7 +84,5 @@ class RequestObject(object):
         if self._can_set_excluded_location(options):
             self.excluded_locations = options['excludedLocations']
 
-    def set_excluded_locations_from_circuit_breaker(self, excluded_locations: List[str]) -> None:
-        if self.excluded_locations:
-            self.excluded_locations.extend(excluded_locations)
-        self.excluded_locations = excluded_locations
+    def set_excluded_locations_from_circuit_breaker(self, excluded_locations: Set[str]) -> None:
+        self.excluded_locations_circuit_breaker = excluded_locations