+++ sdk/cosmos/azure-cosmos/CHANGELOG.md
@@ -1,5 +1,15 @@
 ## Release History
 
+### 4.10.0b5 (Unreleased)
+
+#### Features Added
+
+#### Breaking Changes
+
+#### Bugs Fixed
+
+#### Other Changes
+
 ### 4.10.0b4 (2025-04-01)
 
 #### Bugs Fixed